refactor(Api Folders): The API folder in every module were moved to a diferent location, the route files were also moved
