--- frontend/src/app/admin/establishment/api/routes.js
@@ -1,4 +1,0 @@
-export const getEstablishmentsApiRoute = '/api/establishments';
-export const createEstablishmentsApiRoute = '/api/establishments';
-export const updateEstablishmentsApiRoute = '/api/establishments';
-export const deleteEstablishmentsApiRoute = '/api/establishments';
