+++ frontend/src/app/admin/services/[id]/page.jsx
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from "react";
 import layout from "../../Main.module.css";
 import styles from "./styles.module.css";
 import warning from "../../forms/styles.module.css"
-import { getService } from "../api/services";
+import { getService } from "../../../apiHandlers/adminServices";
 import Buttons from "@/app/components/form/model_buttons/Buttons";
 import { Status } from "@/app/components/form/status/Status";
