+++ frontend/src/app/admin/services/[id]/update/page.jsx
@@ -1,6 +1,6 @@
 "use client";
-import ServiceForm from "../../../forms/ServiceForm";
-import { updateService, getService } from "../../api/services";
+import ServiceForm from "../../../../forms/ServiceForm";
+import { updateService, getService } from "../../../../apiHandlers/adminServices";
 import React, { useEffect, useState } from "react";
 import Layout from "@/app/components/base_layout/Layout";
 import warning from "../../../forms/styles.module.css"
