+++ frontend/src/app/admin/services/bundles/new/page.jsx
@@ -1,7 +1,7 @@
 "use client";
 import { servicesRoute } from "@/app/utils/routes";
-import BundleForm from "../../../forms/BundleForm";
-import { createBundle } from "../../api/services";
+import BundleForm from "../../../../forms/BundleForm";
+import { createBundle } from "../../../../apiHandlers/adminServices";
 import Layout from "@/app/components/base_layout/Layout";
 
 export default function NewService() {
