+++ frontend/src/app/admin/staff/new/page.jsx
@@ -1,5 +1,5 @@
 "use client";
-import EmployeeForm from "../../forms/EmployeeForm";
+import EmployeeForm from "../../../forms/EmployeeForm";
 import { createEmployee } from "../api/employees";
 import Layout from "@/app/components/base_layout/Layout";
