+++ frontend/src/app/routes/adminEstablishment.js
@@ -1,0 +1,1 @@
+export const baseUrl = '/api/establishments/';
