+++ frontend/src/app/routes/adminStaff.js
@@ -1,0 +1,1 @@
+export const baseUrl = "/api/barbers/";
